Disable password autofill for self-hosted login
The password autofill in iOS 12 is really aggressive. The only way I found to make it go away was to set both username and password to content type “one time code”. It seems like anytime there is two text fields, one of which is secure entry, it will offer stored logins.