Remove clearing of siteURLField
This is causing an issue where when there's an XML-RPC error displayed, the text
field will no longer be disabled when the user tries to submit another site URL.

This was first created in cf6b66f6. As far as I can understand, there doesn't
seem to be a critical need to keep this. The reference will still be pointed to
the _current text field_ if the cell is recreated.