Use `Follow Readable Width` for views
Currently the margins for iPad are being manually calculated based on
the view's width. This change removes that calculation and uses the
readable content guide.

Additionally, there were some constraints breaking on the button view
when it was embedded into a view which was larger than itself. I added
some constraints to this view with lower priorities so that it breaks
the correct constraints and does not mess up the buttons.