Revert "Use Xcode 14 in Buildkite CI"
This reverts commit 1d2a541d374fb994b170eac92b40d66b920b642b.

With the Xcode 14 tool-chain, the pod validation fails.

```
error: Cannot code sign because the target does not have an Info.plist file and one is not being generated automatically. Apply an Info.plist file to the target using the INFOPLIST_FILE build setting or generate one automatically by setting the GENERATE_INFOPLIST_FILE build setting to YES (recommended). (in target 'App' from project 'App')
```

Reverting to 13.4.1 in the interest of shipping a new beta. We'll get
back and fix the Xcode 14 issue later.